@@ -3,17 +3,20 @@
   "license": "MIT",
   "scripts": {
     "test": "make test",
-    "build": "make build"
+    "build": "make build",
+    "lint": "make lint",
+    "fix": "make fix"
   },
   "devDependencies": {
     "async": "^1.5.0",
     "babel-core": "^6.13.2",
-    "babel-eslint": "^6.1.2",
+    "babel-eslint": "^7.0.0",
     "babel-plugin-transform-class-properties": "^6.6.0",
     "babel-plugin-transform-flow-strip-types": "^6.3.13",
     "babel-plugin-transform-runtime": "^6.3.13",
     "babel-preset-es2015": "^6.13.2",
     "babel-preset-stage-0": "^6.0.0",
+    "babel-register": "^6.14.0",
     "babel-runtime": "^6.0.0",
     "browserify": "^11.2.0",
     "bundle-collapser": "^1.2.1",
@@ -21,11 +24,11 @@
     "chalk": "1.1.1",
     "codecov.io": "^0.1.6",
     "derequire": "^2.0.2",
-    "eslint": "^2.13.1",
-    "eslint-config-babel": "^1.0.1",
+    "eslint": "^3.7.1",
+    "eslint-config-babel": "^2.0.1",
     "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-flow-vars": "^0.5.0",
-    "flow-bin": "^0.32.0",
+    "eslint-plugin-flowtype": "^2.20.0",
+    "flow-bin": "^0.33.0",
     "gulp": "^3.9.0",
     "gulp-babel": "^6.0.0",
     "gulp-newer": "^1.0.0",
