@@ -1,33 +1,30 @@
-if (process.env.running_under_istanbul) return;
+let readdir        = require("fs-readdir-recursive");
+let helper         = require("babel-helper-fixtures");
+let assert         = require("assert");
+let rimraf         = require("rimraf");
+let outputFileSync = require("output-file-sync");
+let child          = require("child_process");
+let path           = require("path");
+let chai           = require("chai");
+let fs             = require("fs");
+let _              = require("lodash");
 
-var readdir        = require("fs-readdir-recursive");
-var helper         = require("babel-helper-fixtures");
-var assert         = require("assert");
-var rimraf         = require("rimraf");
-var outputFileSync = require("output-file-sync");
-var child          = require("child_process");
-var path           = require("path");
-var chai           = require("chai");
-var fs             = require("fs");
-var pathExists     = require("path-exists");
-var _              = require("lodash");
+let fixtureLoc = path.join(__dirname, "fixtures");
+let tmpLoc = path.join(__dirname, "tmp");
 
-var fixtureLoc = path.join(__dirname, "fixtures");
-var tmpLoc = path.join(__dirname, "tmp");
-
-var presetLocs = [
+let presetLocs = [
   path.join(__dirname, "../../babel-preset-es2015"),
   path.join(__dirname, "../../babel-preset-react")
 ].join(",");
 
-var pluginLocs = [
+let pluginLocs = [
   path.join(__dirname, "/../../babel-plugin-transform-strict-mode"),
   path.join(__dirname, "/../../babel-plugin-transform-es2015-modules-commonjs"),
 ].join(",");
 
-var readDir = function (loc) {
-  var files = {};
-  if (pathExists.sync(loc)) {
+let readDir = function (loc) {
+  let files = {};
+  if (fs.existsSync(loc)) {
     _.each(readdir(loc), function (filename) {
       files[filename] = helper.readFile(path.join(loc, filename));
     });
@@ -35,14 +32,14 @@ var readDir = function (loc) {
   return files;
 };
 
-var saveInFiles = function (files) {
+let saveInFiles = function (files) {
   _.each(files, function (content, filename) {
     outputFileSync(filename, content);
   });
 };
 
-var assertTest = function (stdout, stderr, opts) {
-  var expectStderr = opts.stderr.trim();
+let assertTest = function (stdout, stderr, opts) {
+  let expectStderr = opts.stderr.trim();
   stderr = stderr.trim();
 
   if (opts.stderr) {
@@ -55,7 +52,7 @@ var assertTest = function (stdout, stderr, opts) {
     throw new Error("stderr:\n" + stderr);
   }
 
-  var expectStdout = opts.stdout.trim();
+  let expectStdout = opts.stdout.trim();
   stdout = stdout.trim();
   stdout = stdout.replace(/\\/g, "/");
 
@@ -70,20 +67,19 @@ var assertTest = function (stdout, stderr, opts) {
   }
 
   _.each(opts.outFiles, function (expect, filename) {
-    var actual = helper.readFile(filename);
+    let actual = helper.readFile(filename);
     chai.expect(actual).to.equal(expect, "out-file " + filename);
   });
 };
 
-var buildTest = function (binName, testName, opts) {
-  var binLoc = path.join(__dirname, "../lib", binName);
+let buildTest = function (binName, testName, opts) {
+  let binLoc = path.join(__dirname, "../lib", binName);
 
   return function (callback) {
-    this.timeout(5000);
     clear();
     saveInFiles(opts.inFiles);
 
-    var args = [binLoc];
+    let args = [binLoc];
 
     if (binName !== "babel-external-helpers") {
       args.push("--presets", presetLocs, "--plugins", pluginLocs);
@@ -95,10 +91,10 @@ var buildTest = function (binName, testName, opts) {
 
     args = args.concat(opts.args);
 
-    var spawn = child.spawn(process.execPath, args);
+    let spawn = child.spawn(process.execPath, args);
 
-    var stderr = "";
-    var stdout = "";
+    let stderr = "";
+    let stdout = "";
 
     spawn.stderr.on("data", function (chunk) {
       stderr += chunk;
@@ -109,7 +105,7 @@ var buildTest = function (binName, testName, opts) {
     });
 
     spawn.on("close", function () {
-      var err;
+      let err;
 
       try {
         assertTest(stdout, stderr, opts);
@@ -131,9 +127,9 @@ var buildTest = function (binName, testName, opts) {
   };
 };
 
-var clear = function () {
+let clear = function () {
   process.chdir(__dirname);
-  if (pathExists.sync(tmpLoc)) rimraf.sync(tmpLoc);
+  if (fs.existsSync(tmpLoc)) rimraf.sync(tmpLoc);
   fs.mkdirSync(tmpLoc);
   process.chdir(tmpLoc);
 };
@@ -141,23 +137,23 @@ var clear = function () {
 _.each(fs.readdirSync(fixtureLoc), function (binName) {
   if (binName[0] === ".") return;
 
-  var suiteLoc = path.join(fixtureLoc, binName);
-  suite("bin/" + binName, function () {
+  let suiteLoc = path.join(fixtureLoc, binName);
+  describe("bin/" + binName, function () {
     _.each(fs.readdirSync(suiteLoc), function (testName) {
       if (testName[0] === ".") return;
 
-      var testLoc = path.join(suiteLoc, testName);
+      let testLoc = path.join(suiteLoc, testName);
 
-      var opts = {
+      let opts = {
         args: []
       };
 
-      var optionsLoc = path.join(testLoc, "options.json");
-      if (pathExists.sync(optionsLoc)) _.merge(opts, require(optionsLoc));
+      let optionsLoc = path.join(testLoc, "options.json");
+      if (fs.existsSync(optionsLoc)) _.merge(opts, require(optionsLoc));
 
       _.each(["stdout", "stdin", "stderr"], function (key) {
-        var loc = path.join(testLoc, key + ".txt");
-        if (pathExists.sync(loc)) {
+        let loc = path.join(testLoc, key + ".txt");
+        if (fs.existsSync(loc)) {
           opts[key] = helper.readFile(loc);
         } else {
           opts[key] = opts[key] || "";
@@ -167,13 +163,13 @@ _.each(fs.readdirSync(fixtureLoc), function (binName) {
       opts.outFiles = readDir(path.join(testLoc, "out-files"));
       opts.inFiles  = readDir(path.join(testLoc, "in-files"));
 
-      var babelrcLoc = path.join(testLoc, ".babelrc");
-      if (pathExists.sync(babelrcLoc)) {
+      let babelrcLoc = path.join(testLoc, ".babelrc");
+      if (fs.existsSync(babelrcLoc)) {
         // copy .babelrc file to tmp directory
-        opts.inFiles['.babelrc'] = helper.readFile(babelrcLoc);
+        opts.inFiles[".babelrc"] = helper.readFile(babelrcLoc);
       }
 
-      test(testName, buildTest(binName, testName, opts));
+      it(testName, buildTest(binName, testName, opts));
     });
   });
 });
