@@ -1,9 +1,9 @@
-var assert = require("assert");
-var chalk = require("chalk");
-var codeFrame = require("..");
+let assert = require("assert");
+let chalk = require("chalk");
+let codeFrame = require("..");
 
-suite("babel-code-frame", function () {
-  test("basic usage", function () {
+describe("babel-code-frame", function () {
+  it("basic usage", function () {
     const rawLines = [
       "class Foo {",
       "  constructor()",
@@ -17,7 +17,7 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("optional column number", function () {
+  it("optional column number", function () {
     const rawLines = [
       "class Foo {",
       "  constructor()",
@@ -30,7 +30,7 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("optional column number", function () {
+  it("optional column number", function () {
     const rawLines = [
       "class Foo {",
       "  constructor()",
@@ -43,7 +43,7 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("maximum context lines and padding", function () {
+  it("maximum context lines and padding", function () {
     const rawLines = [
       "/**",
       " * Sums two numbers.",
@@ -68,7 +68,7 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("no unnecessary padding due to one-off errors", function () {
+  it("no unnecessary padding due to one-off errors", function () {
     const rawLines = [
       "/**",
       " * Sums two numbers.",
@@ -93,12 +93,12 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("tabs", function () {
+  it("tabs", function () {
     const rawLines = [
       "\tclass Foo {",
       "\t  \t\t    constructor\t(\t)",
       "\t};",
-    ].join('\n');
+    ].join("\n");
     assert.equal(codeFrame(rawLines, 2, 25), [
       "  1 | \tclass Foo {",
       "> 2 | \t  \t\t    constructor\t(\t)",
@@ -107,7 +107,7 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("opts.highlightCode", function () {
+  it("opts.highlightCode", function () {
     const rawLines = "console.log('babel')";
     const result = codeFrame(rawLines, 1, 9, {highlightCode: true});
     const stripped = chalk.stripColor(result);
@@ -118,8 +118,8 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("opts.linesAbove", function () {
-    var rawLines = [
+  it("opts.linesAbove", function () {
+    let rawLines = [
       "/**",
       " * Sums two numbers.",
       " *",
@@ -142,8 +142,8 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("opts.linesBelow", function () {
-    var rawLines = [
+  it("opts.linesBelow", function () {
+    let rawLines = [
       "/**",
       " * Sums two numbers.",
       " *",
@@ -165,8 +165,8 @@ suite("babel-code-frame", function () {
     ].join("\n"));
   });
 
-  test("opts.linesAbove and opts.linesBelow", function () {
-    var rawLines = [
+  it("opts.linesAbove and opts.linesBelow", function () {
+    let rawLines = [
       "/**",
       " * Sums two numbers.",
       " *",
