@@ -1,11 +1,11 @@
-var browserify = require("browserify");
-var assert = require("assert");
-var path = require("path");
-var vm = require("vm");
+let browserify = require("browserify");
+let assert = require("assert");
+let path = require("path");
+let vm = require("vm");
 
-suite("browserify", function() {
-  test("babel/register may be used without breaking browserify", function(done) {
-    var bundler = browserify(path.join(__dirname, "fixtures/browserify/register.js"));
+describe("browserify", function() {
+  it("babel/register may be used without breaking browserify", function(done) {
+    let bundler = browserify(path.join(__dirname, "fixtures/browserify/register.js"));
 
     bundler.bundle(function(err, bundle) {
       if (err) return done(err);
@@ -14,6 +14,6 @@ suite("browserify", function() {
       // ensure that the code runs without throwing an exception
       vm.runInNewContext("var global = this;\n" + bundle, {});
       done();
-    })
-  })
+    });
+  });
 });
