@@ -1,18 +1,18 @@
-var assert = require("assert");
-var path = require("path");
-var buildConfigChain = require("../lib/transformation/file/options/build-config-chain");
+let assert = require("assert");
+let path = require("path");
+let buildConfigChain = require("../lib/transformation/file/options/build-config-chain");
 
 function fixture() {
-  var args = [__dirname, "fixtures", "config"];
-  for (var i = 0; i < arguments.length; i ++) {
+  let args = [__dirname, "fixtures", "config"];
+  for (let i = 0; i < arguments.length; i ++) {
     args.push(arguments[i]);
   }
   return path.join.apply(path, args);
 }
 
-suite("buildConfigChain", function () {
-  var oldBabelEnv;
-  var oldNodeEnv;
+describe("buildConfigChain", function () {
+  let oldBabelEnv;
+  let oldNodeEnv;
 
   beforeEach(function () {
     oldBabelEnv = process.env.BABEL_ENV;
@@ -27,12 +27,12 @@ suite("buildConfigChain", function () {
     process.env.NODE_ENV = oldNodeEnv;
   });
 
-  test("dir1", function () {
-    var chain = buildConfigChain({
+  it("dir1", function () {
+    let chain = buildConfigChain({
       filename: fixture("dir1", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: [
@@ -76,12 +76,12 @@ suite("buildConfigChain", function () {
     assert.deepEqual(chain, expected);
   });
 
-  test("dir2", function () {
-    var chain = buildConfigChain({
+  it("dir2", function () {
+    let chain = buildConfigChain({
       filename: fixture("dir2", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: [
@@ -115,12 +115,12 @@ suite("buildConfigChain", function () {
     assert.deepEqual(chain, expected);
   });
 
-  test("env - base", function () {
-    var chain = buildConfigChain({
+  it("env - base", function () {
+    let chain = buildConfigChain({
       filename: fixture("env", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: [
@@ -154,14 +154,14 @@ suite("buildConfigChain", function () {
     assert.deepEqual(chain, expected);
   });
 
-  test("env - foo", function () {
+  it("env - foo", function () {
     process.env.NODE_ENV = "foo";
 
-    var chain = buildConfigChain({
+    let chain = buildConfigChain({
       filename: fixture("env", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: [
@@ -205,15 +205,15 @@ suite("buildConfigChain", function () {
     assert.deepEqual(chain, expected);
   });
 
-  test("env - bar", function () {
+  it("env - bar", function () {
     process.env.NODE_ENV = "foo"; // overridden
     process.env.NODE_ENV = "bar";
 
-    var chain = buildConfigChain({
+    let chain = buildConfigChain({
       filename: fixture("env", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: [
@@ -258,14 +258,14 @@ suite("buildConfigChain", function () {
   });
 
 
-  test("env - foo", function () {
+  it("env - foo", function () {
     process.env.NODE_ENV = "foo";
 
-    var chain = buildConfigChain({
+    let chain = buildConfigChain({
       filename: fixture("pkg", "src.js")
     });
 
-    var expected = [
+    let expected = [
       {
         options: {
           plugins: ["pkg-plugin"]
