@@ -1,10 +1,10 @@
-var a = 1;
-var b = 2;
-var c = 3;
-var d = 4;
-var _e$f = { e: 5, f: 6 };
-var e = _e$f.e;
-var f = _e$f.f;
-var _a$b = { a: 7, b: 8 };
-var g = _a$b.a;
-var h = _a$b.b;
+var a = 1,
+    b = 2;
+var c = 3,
+    d = 4;
+var _e$f = { e: 5, f: 6 },
+    e = _e$f.e,
+    f = _e$f.f;
+var _a$b = { a: 7, b: 8 },
+    g = _a$b.a,
+    h = _a$b.b;
