@@ -1,5 +1,5 @@
 ({
-  *g() {
+  g() {
     var _this = this;
 
     return babelHelpers.asyncGenerator.wrap(function* () {
