@@ -1,5 +1,5 @@
 class C {
-  static *g() {
+  static g() {
     var _this = this;
 
     return babelHelpers.asyncGenerator.wrap(function* () {
