@@ -1,36 +1,36 @@
-var a = 1;
-var b = 2;
-var a = 1;
-var b = 2;
-var a = 1;
-var b = 2;
-var c = [3, 4];
-var a = 1;
-var b = 2;
-var c = [3, 4];
-var _ref = [1, 2, 3];
-var a = _ref[0];
-var b = _ref[1];
-var _ref2 = [1, 2, 3];
-var a = _ref2[0];
-var b = _ref2[1];
-var _ref3 = [a, b];
-var a = _ref3[0];
-var b = _ref3[1];
+var a = 1,
+    b = 2;
+var a = 1,
+    b = 2;
+var a = 1,
+    b = 2,
+    c = [3, 4];
+var a = 1,
+    b = 2,
+    c = [3, 4];
+var _ref = [1, 2, 3],
+    a = _ref[0],
+    b = _ref[1];
+var _ref2 = [1, 2, 3],
+    a = _ref2[0],
+    b = _ref2[1];
+var _ref3 = [a, b],
+    a = _ref3[0],
+    b = _ref3[1];
 var _ref4 = [a[1], a[0]];
 a[0] = _ref4[0];
 a[1] = _ref4[1];
 
-var _ref5 = [].concat(babelHelpers.toConsumableArray(foo), [bar]);
+var _ref5 = [].concat(babelHelpers.toConsumableArray(foo), [bar]),
+    a = _ref5[0],
+    b = _ref5[1];
 
-var a = _ref5[0];
-var b = _ref5[1];
-var _ref6 = [foo(), bar];
-var a = _ref6[0];
-var b = _ref6[1];
-var _ref7 = [clazz.foo(), bar];
-var a = _ref7[0];
-var b = _ref7[1];
-var _ref8 = [clazz.foo, bar];
-var a = _ref8[0];
-var b = _ref8[1];
+var _ref6 = [foo(), bar],
+    a = _ref6[0],
+    b = _ref6[1];
+var _ref7 = [clazz.foo(), bar],
+    a = _ref7[0],
+    b = _ref7[1];
+var _ref8 = [clazz.foo, bar],
+    a = _ref8[0],
+    b = _ref8[1];
