@@ -1,7 +1,7 @@
-var a = "hello";
-var _ref = [", ", "junk"];
-var b = _ref[0];
-var c = "world";
+var a = "hello",
+    _ref = [", ", "junk"],
+    b = _ref[0],
+    c = "world";
 a = "hello";
 var _ref2 = [", ", "junk"];
 b = _ref2[0];
