@@ -1,11 +1,7 @@
-var _ref = ["foo", "hello", [", ", "junk"], ["world"]];
-var a = _ref[1];
-
-var _ref$ = babelHelpers.slicedToArray(_ref[2], 1);
-
-var b = _ref$[0];
-
-var _ref$2 = babelHelpers.slicedToArray(_ref[3], 1);
-
-var c = _ref$2[0];
-var d = _ref[4];
+var _ref = ["foo", "hello", [", ", "junk"], ["world"]],
+    a = _ref[1],
+    _ref$ = babelHelpers.slicedToArray(_ref[2], 1),
+    b = _ref$[0],
+    _ref$2 = babelHelpers.slicedToArray(_ref[3], 1),
+    c = _ref$2[0],
+    d = _ref[4];
