@@ -1,6 +1,6 @@
 var _ref = {};
-var a = _ref.a;
-var b = _ref.b;
+var a = _ref.a,
+    b = _ref.b;
 var _ref$c = _ref.c;
 var d = _ref$c.d;
 var _ref$c$e$f = _ref$c.e.f;
