@@ -1,3 +1,3 @@
 var z = [];
-var x = z[0];
-var y = z.slice(1);
+var x = z[0],
+    y = z.slice(1);
