@@ -1,11 +1,8 @@
 var rect = {};
 
-var _rect$topLeft = babelHelpers.slicedToArray(rect.topLeft, 2);
-
-var x1 = _rect$topLeft[0];
-var y1 = _rect$topLeft[1];
-
-var _rect$bottomRight = babelHelpers.slicedToArray(rect.bottomRight, 2);
-
-var x2 = _rect$bottomRight[0];
-var y2 = _rect$bottomRight[1];
+var _rect$topLeft = babelHelpers.slicedToArray(rect.topLeft, 2),
+    x1 = _rect$topLeft[0],
+    y1 = _rect$topLeft[1],
+    _rect$bottomRight = babelHelpers.slicedToArray(rect.bottomRight, 2),
+    x2 = _rect$bottomRight[0],
+    y2 = _rect$bottomRight[1];
