@@ -1,4 +1,4 @@
 var coords = [1, 2];
-var x = coords.x;
-var y = coords.y;
-var foo = "bar";
+var x = coords.x,
+    y = coords.y,
+    foo = "bar";
