@@ -1,10 +1,10 @@
 var rect = {};
-var _rect$topLeft = rect.topLeft;
-var x1 = _rect$topLeft.x;
-var y1 = _rect$topLeft.y;
-var _rect$bottomRight = rect.bottomRight;
-var x2 = _rect$bottomRight.x;
-var y2 = _rect$bottomRight.y;
-var _ref = [0, 1, 2, 3, 4, 5, 6];
-var foo = _ref[3];
-var bar = _ref[5];
+var _rect$topLeft = rect.topLeft,
+    x1 = _rect$topLeft.x,
+    y1 = _rect$topLeft.y,
+    _rect$bottomRight = rect.bottomRight,
+    x2 = _rect$bottomRight.x,
+    y2 = _rect$bottomRight.y;
+var _ref = [0, 1, 2, 3, 4, 5, 6],
+    foo = _ref[3],
+    bar = _ref[5];
