@@ -1,3 +1,3 @@
 var coords = [1, 2];
-var x = coords.x;
-var y = coords.y;
+var x = coords.x,
+    y = coords.y;
