@@ -8,8 +8,7 @@ var _foo2 = _interopRequireDefault(_foo);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-var _bar = bar;
-
-var _bar2 = _slicedToArray(_bar, 1);
+var _bar = bar,
+    _bar2 = _slicedToArray(_bar, 1);
 
 const x = _bar2[0];
