@@ -1,5 +1,5 @@
 class X {
   foo = 2;
   bar /*: number*/ = 3;
-  baz /*: ?string*/;
+  /*:: baz: ?string*/
 }
