@@ -2,7 +2,6 @@ import path from "path";
 import fs from "fs";
 import { sync as mkdirpSync } from "mkdirp";
 import homeOrTmp from "home-or-tmp";
-import pathExists from "path-exists";
 
 const FILENAME = process.env.BABEL_CACHE_PATH || path.join(homeOrTmp, ".babel.json");
 let data = {};
@@ -37,7 +36,7 @@ export function load() {
   process.on("exit", save);
   process.nextTick(save);
 
-  if (!pathExists.sync(FILENAME)) return;
+  if (!fs.existsSync(FILENAME)) return;
 
   try {
     data = JSON.parse(fs.readFileSync(FILENAME));
