@@ -1,10 +1,10 @@
-var traverse = require("../lib").default;
-var assert = require("assert");
-var parse = require("babylon").parse;
+let traverse = require("../lib").default;
+let assert = require("assert");
+let parse = require("babylon").parse;
 
 function getPath(code) {
-  var ast = parse(code);
-  var path;
+  let ast = parse(code);
+  let path;
   traverse(ast, {
     Program: function (_path) {
       path = _path;
@@ -14,16 +14,16 @@ function getPath(code) {
   return path;
 }
 
-suite("evaluation", function () {
-  suite("evaluateTruthy", function () {
-    test("it should work with null", function () {
+describe("evaluation", function () {
+  describe("evaluateTruthy", function () {
+    it("it should work with null", function () {
       assert.strictEqual(
         getPath("false || a.length === 0;").get("body")[0].evaluateTruthy(),
         undefined
       );
     });
 
-    test("it should not mistake lack of confidence for falsy", function () {
+    it("it should not mistake lack of confidence for falsy", function () {
       assert.strictEqual(
         getPath("foo || 'bar'").get("body")[0].evaluate().value,
         undefined
@@ -31,35 +31,35 @@ suite("evaluation", function () {
     });
   });
 
-  test("should bail out on recursive evaluation", function () {
+  it("should bail out on recursive evaluation", function () {
     assert.strictEqual(
       getPath("function fn(a) { var g = a ? 1 : 2, a = g * this.foo; }").get("body.0.body.body.0.declarations.1.init").evaluate().confident,
       false
     );
   });
 
-  test("should work with repeated, indeterminate identifiers", function () {
+  it("should work with repeated, indeterminate identifiers", function () {
     assert.strictEqual(
       getPath("var num = foo(); (num > 0 && num < 100);").get("body")[1].evaluateTruthy(),
       undefined
     );
   });
 
-  test("should work with repeated, determinate identifiers", function () {
+  it("should work with repeated, determinate identifiers", function () {
     assert.strictEqual(
       getPath("var num = 5; (num > 0 && num < 100);").get("body")[1].evaluateTruthy(),
       true
     );
   });
 
-  test("should deopt when var is redeclared in the same scope", function () {
+  it("should deopt when var is redeclared in the same scope", function () {
     assert.strictEqual(
       getPath("var x = 2; var y = x + 2; { var x = 3 }").get("body.1.declarations.0.init").evaluate().confident,
       false
     );
   });
 
-  test("it should not deopt vars in different scope", function () {
+  it("it should not deopt vars in different scope", function () {
     const input = "var a = 5; function x() { var a = 5; var b = a + 1; } var b = a + 2";
     assert.strictEqual(
       getPath(input).get("body.1.body.body.1.declarations.0.init").evaluate().value,
@@ -71,7 +71,7 @@ suite("evaluation", function () {
     );
   });
 
-  test("it should not deopt let/const inside blocks", function () {
+  it("it should not deopt let/const inside blocks", function () {
     assert.strictEqual(
       getPath("let x = 5; { let x = 1; } let y = x + 5").get("body.2.declarations.0.init").evaluate().value,
       10
