@@ -1,9 +1,9 @@
-var traverse = require("../lib").default;
-var assert   = require("assert");
-var _        = require("lodash");
+let traverse = require("../lib").default;
+let assert   = require("assert");
+let _        = require("lodash");
 
-suite("traverse", function () {
-  var ast = {
+describe("traverse", function () {
+  let ast = {
     type: "Program",
     body: [
       {
@@ -50,14 +50,14 @@ suite("traverse", function () {
     ]
   };
 
-  var body = ast.body;
+  let body = ast.body;
 
-  test("traverse replace", function () {
-    var replacement = {
+  it("traverse replace", function () {
+    let replacement = {
       type: "StringLiteral",
       value: "foo"
     };
-    var ast2 = _.cloneDeep(ast);
+    let ast2 = _.cloneDeep(ast);
 
     traverse(ast2, {
       enter: function (path) {
@@ -68,13 +68,13 @@ suite("traverse", function () {
     assert.equal(ast2.body[1].expression.left.object, replacement);
   });
 
-  test("traverse", function () {
-    var expect = [
+  it("traverse", function () {
+    let expect = [
       body[0], body[0].declarations[0], body[0].declarations[0].id, body[0].declarations[0].init,
       body[1], body[1].expression, body[1].expression.left, body[1].expression.left.object, body[1].expression.left.property, body[1].expression.right
     ];
 
-    var actual = [];
+    let actual = [];
 
     traverse(ast, {
       enter: function (path) {
@@ -85,7 +85,7 @@ suite("traverse", function () {
     assert.deepEqual(actual, expect);
   });
 
-  test("traverse falsy parent", function () {
+  it("traverse falsy parent", function () {
     traverse(null, {
       enter: function () {
         throw new Error("should not be ran");
@@ -93,13 +93,13 @@ suite("traverse", function () {
     });
   });
 
-  test("traverse blacklistTypes", function () {
-    var expect = [
+  it("traverse blacklistTypes", function () {
+    let expect = [
       body[0], body[0].declarations[0], body[0].declarations[0].id, body[0].declarations[0].init,
       body[1], body[1].expression, body[1].expression.right
     ];
 
-    var actual = [];
+    let actual = [];
 
     traverse(ast, {
       blacklist: ["MemberExpression"],
@@ -111,7 +111,7 @@ suite("traverse", function () {
     assert.deepEqual(actual, expect);
   });
 
-  test("hasType", function () {
+  it("hasType", function () {
     assert.ok(traverse.hasType(ast, null, "ThisExpression"));
     assert.ok(!traverse.hasType(ast, null, "ThisExpression", ["AssignmentExpression"]));
 
@@ -124,8 +124,8 @@ suite("traverse", function () {
     assert.ok(!traverse.hasType(ast, null, "ArrowFunctionExpression"));
   });
 
-  test("clearCache", function () {
-    var paths = [];
+  it("clearCache", function () {
+    let paths = [];
     traverse(ast, {
       enter: function (path) {
         paths.push(path);
@@ -134,7 +134,7 @@ suite("traverse", function () {
 
     traverse.clearCache();
 
-    var paths2 = [];
+    let paths2 = [];
     traverse(ast, {
       enter: function (path) {
         paths2.push(path);
